# openapi.yaml
openapi: 3.0.3
info:
  title: My Finances API
  version: 1.0.0
  description: Personal finance tracker — Plaid sync, investments, expenses, income, assets.
servers:
  - url: http://127.0.0.1:8000/api
paths:
...
  /investments/holdings:
    get:
      summary: All holdings (optionally filtered by account)
      tags: [investments]
      parameters:
        - name: account_id
          in: query
          schema: { type: string }
...
  /investments/history:
    get:
      summary: Portfolio value over time (daily snapshots)
      tags: [investments]
      parameters:
        - name: months
          in: query
          schema: { type: integer, default: 12 }
  /investments/transactions:
    get:
      summary: Investment transactions (buys, sells, dividends, fees)
      tags: [investments]
      parameters:
        - name: start
          in: query
          required: true
          schema: { type: string, format: date }
        - name: end
          in: query
          required: true
          schema: { type: string, format: date }
        - name: account_id
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string }
...